Fix cross-language sign2/sign1 verification failure: timestamp msgpack type mismatch
Real Go client <-> TS server handshake: the TS server accepted the
client's INIT and produced a STEP1 with no complaints, but the Go
client rejected that STEP1 with "failed to verify sign2: invalid
signature" -- Go's re-derivation of the unsigned bytes for checking
the server's sign2 didn't byte-match what TS actually signed.

Root cause (confirmed by byte-level decode): Challenge.Timestamp was
Go int64, which vmihailenco/msgpack encodes with the signed msgpack
type (0xd3). TS's @msgpack/msgpack encodes the same non-negative
Date.now() value with the unsigned type (0xcf). Same numeric value,
different wire bytes, so every byte-exact signature check across the
two languages failed deterministically -- this wasn't data-dependent,
it would fail for every handshake.

Fix: Challenge.Timestamp and orderedChallengeData.Timestamp are now
uint64 (timestamps are never negative, so no semantic change), with
explicit casts at the two remaining boundaries that stay int64:
ChallengerState.Timestamp (internal, not serialized) and
validateTimestamp's parameter. Verified the encoded field now carries
the 0xcf tag instead of 0xd3 for the same value.

+++ go/pkg/challenger/challenger.go
@@ -61,7 +61,7 @@ func (c *Challenger) Init(protocol, service string) (*Challenge, error) {
 		Version:   c.options.Version,
 		Protocol:  protocol,
 		Service:   service,
-		Timestamp: c.state.Timestamp,
+		Timestamp: uint64(c.state.Timestamp),
 		PK1:       c.vaultysID.ToBytes(),
 		Nonce:     nonce,
 		Metadata: ChallengeMetadata{
@@ -86,7 +86,7 @@ func (c *Challenger) Step1(initChallenge *Challenge) (*Challenge, error) {
 	}
 
 	// Check timestamp
-	if err := c.validateTimestamp(initChallenge.Timestamp); err != nil {
+	if err := c.validateTimestamp(int64(initChallenge.Timestamp)); err != nil {
 		return nil, err
 	}
 
@@ -95,7 +95,7 @@ func (c *Challenger) Step1(initChallenge *Challenge) (*Challenge, error) {
 	c.state.RemoteNonce = initChallenge.Nonce
 	c.state.Protocol = initChallenge.Protocol
 	c.state.Service = initChallenge.Service
-	c.state.Timestamp = initChallenge.Timestamp
+	c.state.Timestamp = int64(initChallenge.Timestamp)
 
 	// Generate our 16-byte nonce
 	ourNonce, err := crypto.RandomBytes(16)
@@ -460,7 +460,7 @@ type orderedChallengeData struct {
 	Version   uint8             `msgpack:"version"`
 	Protocol  string            `msgpack:"protocol"`
 	Service   string            `msgpack:"service"`
-	Timestamp int64             `msgpack:"timestamp"`
+	Timestamp uint64            `msgpack:"timestamp"`
 	PK1       []byte            `msgpack:"pk1"`
 	PK2       []byte            `msgpack:"pk2"`
 	Nonce     []byte            `msgpack:"nonce"`
